Make textarea match height of English content
Uses the `lh` CSS length unit to make the textarea at least 5 lines
tall, then uses flexbox to make it grow if the rest of the row is longer
than that.

`lh` units are relatively new but they're supported in all of our target
browsers, and they handily equate to 1 'line' since an `lh` is equal to
the computed value of line-height.

Also uses `field-sizing: content` to make the textarea scale with the
length of the user's input. This is a new CSS feature currently
supported in Chromium-based browsers, and added in the latest version of
Safari, so it doesn't work in current versions of Firefox. However
Mozilla have announced intent to ship it and the Bugzilla ticket for
implementing it has been assigned:
https://bugzilla.mozilla.org/show_bug.cgi?id=1832409

I think it is reasonable to treat this as an enhancement for the
browsers that support it - it's probably not worth implementing
equivalent JS behaviour for the short time it's unsupported by Firefox.