Add new none of the above option for selection questions
Make it so that users can select "Yes, and let people provide a
different answer" for "Should the list include an option for 'None of
the above'?". Only show this option if the feature flag is enabled.

If either "Yes" or "No" are selected, delete the existing
`none_of_the_above_question` configuration from the DraftQuestion if
it exists.