Update metrics content
Realised there were a couple more tweaks needed here for accuracy. 

Have removed "following for the past 7 days" as people will be able to get metrics for a longer period than that. 

Made the bullets lower case for style. 

Removed "no metrics are available" as there may still be metrics that are older than 7 days - in which case they will get a link to download the CSV.