Show the none of the above question on the edit question page
If "Yes, and let people provide a different answer" was selected for
"Should the list include an option for ‘None of the above’?", show
the question that will be asked on the edit question page with a link
to change it.

If the question is optional, show " (optional)" after the question.