Ensure correct none of the above option is selected when going back
When the back button is clicked on the page to configure the none of
the above question, we want to make sure the correct option for
whether to include none of the above is selected. To acheive this,
when the selection options page is submitted, as well as setting
`is_optional` based on the selected option, also set
`none_of_the_above_question` to be an empty array in the
answer_settings so we can use this to determine that the "Yes, and let
people provide a different answer" option was selected.