Add value field when making selection options
To support routing in Welsh forms, we are adding a new field to the
selection options.

The field, called 'value', will be used in the runner as the single
value when make decisions for routing. It will always contain the
English name for the option.

For most forms which will only be in English, this will be the same as
the name field. But for Welsh forms, the name will be in Welsh and the
value will be in English.

This commit makes sure that when selection options are created, they
have both fields.