CloudWatch service requests both metric namespaces
Because we have recently changed the CloudWatch namespace under which we
store our metrics, we have data that is stored under both for forms that
existed prior to the switch. This means that we need to request data
under both the old and new namespaces, and combine them.

The older data is prioritised over the newer data when there is a
conflict, as the old data would have been collected from the start of
the day that the switch took place - whereas the new data may only have
been collected partway through the day. There should only be one day in
a form's metric history that has potentially differing data, as days
either side will have metrics under both the old and new namespace.