Add task add_value_to_selection_options
Add a new rake task to add "value" attribute to selection_options in
FormDocuments.

This is a data migration which will query for all form documents which
have a step with an answer_type of :selection.

It then loops through each step and adds a value attribute to each
selection option.