Refactor updating question
To be more consistent with how we usually do updates, move updating
the page when submitting the edit question page to the `submit` method
on the QuestionInput.

Previously we would have wanted this to be in the controller as it was
calling forms-api using ActiveResource. But now the Page is updated in
the local database instead.

Improve the tests around this.