Clear empty none_of_the_above_question when saving question
This is to handle the case when a user clicks the "Change" link on the
edit question page for a selection question and changes their answer
for "Should the list include an option for ‘None of the above’?" to
"Yes, and let people provide a different answer" but then clicks back
without entering a question text.

In this case, when they come to save their question, remove the blank
hash for none_of_the_above_question in the answer_settings. Form
fillers will be shown a "None of the above" option with no question in
this case.