Set metrics lookback period to 14 months
We noticed an issue where the 15 month lookback would ocassionally look
back further than the available recorded metrics data, which resulted in
unusual timing windows being retrieved.

CloudWatch keeps data for 455 days, but this can mean that 15 months ago
can sometimes fall outside that range. This resulted in erroneous
data being returned. We've opted instead for 14 months, which makes it
clearer in terms of content and what data is expected, and avoids
the possiblity of attempting to retrieve data from out of range.