Fix incorrect column width
A column width of 25% had been set on the row-spanning header cells in
the translation table. However, the same width had not been applied to
the empty header cell at the top left of the table, so the width was not
applied.

This commit adds a class to the empty cell and applies the width to it.