Only test that one user role is forbidden
It is very expensive to test that both user roles that are not a super
admin are forbidden from accessing reports. We only really need to
test that the policy is applied - and for most controllers we only
test this for one action so even this is probably overkill.

This reduced runtime from 46 seconds to 35 seconds for this spec file.