Combine examples in reports_controller_spec for speedup
Having a single example per expectation is standard for rspec. This
makes sense to help readability for unit tests where this is
inexpensive. However, for request tests there is a lot of overhead
to run the set-up per it block. Combine expectatitons into fewer
examples to speed up the test suite.

This reduces the time taken for this spec file from 35s to 20s.